pyproject.toml
[project]
name = "twm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.3",
]
